[Feature] Support for automatic increment of numeric values
Added support for automatic increment of values for both standard data and meta-data, on a per-data-type basis. If the same data-type is used both as meta-data and standard data their incrementing values are kept separate, unless the meta-data prefix is the same as the standard data identifier.
Added tests.
Updated ReadMe to explain how the feature works.