[REFACTOR] Add explicit typeless-data classes
Add a set of typeless-data classes to simplify handing typeless-data. Base abstract Typeless with generic sub-class, then TypelessNames to hold the field names, TypelessValues to hold a row of values, finally Typelessdata to hold a set of field names and a list of values.
Refactor the Converter class to use those instead of the explicit string[] types.