[FEATURE] Conversion support
Add ConversionType and ConversionValue model classes.
Rename ConversionHelper to ConverterHelper. This will allow a new ConversionHelpter static class to target the type and value conversion functionality. While the ConverterHelpter targets the Converter class.

[REFACTOR] Add helpers

Add ConfigHelper and MetadataHelper.

[FEATURE] Add ConversionHandler and ConversionHelper classes

Add the ConversionHandler class to parse the lines for conversion types and values, hold them for future use and manage the conversions.
Add the ConversionHelper class to support the construction of arrays of ConversionType and ConvertionValue objects.
Update Options to specify the prefixes and whether conversions are enabled or not.
Update ConfigHandler to support Conversion prefixes and enabled flags.

[TEST] Add tests

Add test for ConversionType and ConversionValue model objects; Handler constructor, options including defalts, enable / disable conversion, and prefixes; Handler conversion methods.

[FEATURE] Add use of conversion functionality to converter

Update Converter to use the ConversionHandler for both types and values.
Inject the ConversionHandler into the ValueConverter so it can request the value conversion when converting a string value.
Inject the ConversionHandler into the MetadataHandler so it can pass it to the ValueConverter it uses.
Convert the type name when generating the dictionary of Typeless values.

[TESTS] Add tests

Add tests to confirm that typed results have values conversions applied if enabled, but not if disabled. Ditto typeless results. And that typeless results have typename conversions applied if enabled but not if disabled.

[REFACTOR] Split serialization and deserialisation out to their own processors

Created SerializationProcessor and DeserializationProcessor. Separated the orchestration of the handlers from the serialization / deserialization process. The Converter class now only does the orchestration.