[TEST]{FIX] Various real-world bug fixes and tests to exercise them
Add retrieval of only instance properties when retrieving PropertyAndAttributePair objects.
Add filtering out of types that aren't present in the CSV when constructing typeless data prior to converting to typed data.
Add method to check to see if the text looks like it could be CSV data.
Add PropertyGroup elements to support the generation of a NuGet package.
Add tests.