[FEATURE][TESTS] Fluent serialization
Add a public interface and an internal class to handle fluent serialization, and a related method to the Converter to return a suitably initialised object.
Renamed OptionMetadata to OptionTypedMetadata and OptionMetadataDictionary to OptionTypelessMetadata to better reflect their function.
Add Option extension methods to retrieve OptionMetadata according to prefix or Type, and to retrieve the property names from a metadata array relating to its type.
Added dictionary keys to the converter to expose the keys used for the global and typed configuration dictionary.
Added tests.

[REFACTOR] Refactor SerializationProcessor and Serialization

Refactor SerializationProcessor and Serialization in preparation to using the latter as the funtionality of the former.
Add OptionsExtensions for OptionType and OptionMember.
Add TypeExtensions for a property's declaring type's depth.