[REFACTOR} tweaks to fluent serialization interface methods
Provide three methods for typed metadata and typeless metadata; One where the data-type is specified by a TypeParameter, one by a Type and one by a string containing the type name.
Change the Comment fluent serialization method to accept a params string[] so an array of comment text can be supplied.
For OptionTypedMetadata change T to TMetadata to better describe its function.